+++ src/@types/settings.d.ts
@@ -37,7 +37,8 @@ export interface Settings {
         | "ac_white_plug"
         | "ac_white_plug_hollow"
         | "ac_black_plug"
-        | "ac_black_plug_hollow";
+        | "ac_black_plug_hollow"
+        | "default"; // old configs DON'T USE
     doneSetup: boolean;
     clientName: string;
     smoothScroll: boolean;
