+++ src/setup/setup.css
@@ -101,14 +101,6 @@ div {
     font-size: 100%;
     vertical-align: baseline;
 }
-#logo {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    content: url("legcord://assets/ac_logo.svg");
-    width: 292px;
-}
 
 /* }}} */
