+++ src/setup/setup.html
@@ -20,7 +20,6 @@
         </p>
       </div>
       <div id="setup">
-        <div id="logo" class="hidden"></div>
         <div id="page1" class="hidden">
           <p id="setup_question1">Welcome to Legcord Setup</p>
           <div id="buttons">
@@ -105,9 +104,6 @@
               };
               // Accessors {{{
               let options = {};
-
-              let logo = document.getElementById("logo");
-              logo.classList.remove("hidden");
 
               let page1 = document.getElementById("page1");
               page1.classList.remove("hidden");
